add general thread `Worker` class and move map area creation there + add a couple of timers to measure different stages of the rendering process
Worker-creation seems to work great. The requested map area might not
be there in time but there's virtually no time consumed with the map
update. The TiledMapNode.render() method does not seem to consume
relevant time at all but there is a noticeable lag in rendering and
the PipelineStage render time is pretty bad every now and then.
The relevant gl operations involved seem to be the TiledMapNode
multisample-downsampling or the texture-binding. Not clear why..